refactor: rename common package to idmap for better clarity (#55)
- Renamed directory internal/core/common/ to internal/core/idmap/
- Updated package declaration from 'common' to 'idmap'
- Updated all imports across the codebase (8 files)
- Removed //nolint:revive directive as it's no longer needed

This change addresses the linter warning about meaningless package names
and provides a more descriptive name that reflects the actual purpose
of the package (ID mapping functionality).
