+++ internal/cli/commands/helpers.go
@@ -3,15 +3,15 @@ package commands
 import (
 	"fmt"
 
-	"github.com/aki/amux/internal/core/common"
 	"github.com/aki/amux/internal/core/config"
+	"github.com/aki/amux/internal/core/idmap"
 	"github.com/aki/amux/internal/core/mailbox"
 	"github.com/aki/amux/internal/core/session"
 	"github.com/aki/amux/internal/core/workspace"
 )
 
 // createSessionManager is a helper to create a session manager with all dependencies
-func createSessionManager(configManager *config.Manager, wsManager *workspace.Manager, idMapper *common.IDMapper) (*session.Manager, error) {
+func createSessionManager(configManager *config.Manager, wsManager *workspace.Manager, idMapper *idmap.IDMapper) (*session.Manager, error) {
 	store, err := session.NewFileStore(configManager.GetAmuxDir())
 	if err != nil {
 		return nil, fmt.Errorf("failed to create session store: %w", err)
